+++ Pandora/Paradigm/Structure/Ability/Morphable.hs
@@ -23,7 +23,7 @@ premorph = extract . run
 
 data Walk a = Preorder a | Inorder a | Postorder a | Levelorder a
 
-data Morph a = Rotate a | Into a | Insert a | Push a
+data Morph a = Rotate a | Into a | Insert a | Push a | Pop a
 
 rotate :: forall f t . Morphable (Rotate f) t => t ~> Morphing (Rotate f) t
 rotate = morphing . TU . Tag @(Rotate f)